add docker com mysql

+++ src/main/java/com/impacta/microservices/investimento/repository/InvestimentoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InvestimentoRepository extends JpaRepository<Investimento,Long> {
+public interface InvestimentoRepository extends JpaRepository<Investimento,Integer> {
 
     Investimento findByContaId(Integer contaId);
 }
